@@ -6,31 +6,48 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.2
-  flutter_native_splash: ^2.1.6
-  flutter_launcher_icons: ^0.9.2
+  #/ ADS
+  google_mobile_ads: ^2.2.0
+  #/ APP NAME & ICON
   rename: ^2.0.1
+  flutter_launcher_icons: ^0.11.0
+  flutter_native_splash: ^2.1.6
+  #/ CONTROLLER
   get: ^4.6.1
+  #/ DATABASE
+  sqflite: ^2.2.0+3
+  sqflite_common_ffi_web: ^0.3.1
+  #/ PATH
+  path_provider: ^2.0.11
+  url_launcher: ^6.1.7
+  # audioplayers: ^1.1.1
+  #/ JAPANESE FUNCTION
+  kana_kit: ^2.0.0
+  unofficial_jisho_api: ^3.0.0
+  #/ STYLE
+  google_fonts: ^3.0.1
+  auto_scroll_text: ^0.0.6
+  lottie: ^2.1.0
+  animate_do: ^3.0.2
+  shake_animation_widget: ^3.0.4
+  text_transformation_animation: ^0.1.4
+  audioplayers: ^1.1.1
+  linkfy_text: ^1.1.5
+  flutter_keyboard_visibility: ^5.4.0
   get_storage: ^2.0.3
-  firebase_core: ^1.14.1
-  firebase_storage: ^10.2.12
-  path_provider: ^2.0.9
-  store_redirect: ^2.0.1
-  expandable: ^5.0.1
-  url_launcher: ^6.1.0
-  flutter_email_sender: ^5.1.0
-  url_strategy: ^0.2.0
-  google_mobile_ads: ^1.3.0
+  desktop_window: ^0.4.0
+  flutter_markdown: ^0.6.13
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^2.0.1
 
 flutter_native_splash:
   color: "#BC002D"
@@ -40,10 +57,27 @@ flutter_native_splash:
 flutter_icons:
   android: true
   image_path: "assets/logo.png"
+  web:
+    generate: true
+    image_path: "assets/logo.png"
+    background_color: "#BC002D"
+    theme_color: "#BC002D"
+  windows:
+    generate: true
+    image_path: "assets/logo.png"
+    icon_size: 48 # min:48, max:256, default: 48
 
 flutter:
   uses-material-design: true
   assets:
-    - assets/
-    - assets/help/
-    - assets/tips/
+  #/ LOGO
+  - assets/logo.png
+  #/ DICTIONARY
+  - assets/wordlist_database/
+  #/ SCREEN EFFECT
+  - assets/Graphics/
+  #/ SOUND
+  - assets/Sound Effect/
+  #/ ABOUT APP
+  - assets/About App.md
+
